requirements.txt
fastapi
pytest
factory_boy